Remove uint cast when JIT can prove never negative

+++ src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/ReadOnlyTensorSpan.cs
@@ -92,7 +92,7 @@ namespace System.Numerics.Tensors
             if (Environment.Is64BitProcess)
             {
                 // See comment in Span<T>.Slice for how this works.
-                if ((ulong)(uint)start + (ulong)(uint)maxElements >= (ulong)(uint)array.Length && array.Length != 0)
+                if ((ulong)(uint)start + (ulong)(uint)maxElements >= (ulong)array.Length && array.Length != 0)
                     ThrowHelper.ThrowArgument_InvalidStridesAndLengths();
             }
             else
@@ -187,7 +187,7 @@ namespace System.Numerics.Tensors
             if (Environment.Is64BitProcess)
             {
                 // See comment in Span<T>.Slice for how this works.
-                if ((ulong)(uint)startOffset + (ulong)(uint)maxElements >= (ulong)(uint)array.Length && array.Length != 0)
+                if ((ulong)(uint)startOffset + (ulong)(uint)maxElements >= (ulong)array.Length && array.Length != 0)
                     ThrowHelper.ThrowArgumentOutOfRangeException();
             }
             else
@@ -235,7 +235,7 @@ namespace System.Numerics.Tensors
             if (Environment.Is64BitProcess)
             {
                 // See comment in Span<T>.Slice for how this works.
-                if ((ulong)(uint)startOffset + (ulong)(uint)maxElements > (ulong)(uint)array.Length)
+                if ((ulong)(uint)startOffset + (ulong)(uint)maxElements > (ulong)array.Length)
                     ThrowHelper.ThrowArgumentOutOfRangeException();
             }
             else
